+++ src/main/java/com/sparta/bookflex/domain/auth/dto/RefreshResDto.java
@@ -1,4 +1,4 @@
-package com.sparta.bookflex.domain.user.dto;
+package com.sparta.bookflex.domain.auth.dto;
 
 import lombok.Getter;
