+++ src/main/java/com/sparta/bookflex/domain/reveiw/service/ReviewService.java
@@ -11,11 +11,10 @@ import com.sparta.bookflex.domain.reveiw.dto.ReviewResponseDto;
 import com.sparta.bookflex.domain.reveiw.entity.Review;
 import com.sparta.bookflex.domain.reveiw.repository.ReviewCustomRepositoryImpl;
 import com.sparta.bookflex.domain.reveiw.repository.ReviewRepository;
-import com.sparta.bookflex.domain.sale.entity.Sale;
 import com.sparta.bookflex.domain.sale.repository.SaleRepository;
 import com.sparta.bookflex.domain.user.entity.User;
 import com.sparta.bookflex.domain.user.enums.RoleType;
-import com.sparta.bookflex.domain.user.service.AuthService;
+import com.sparta.bookflex.domain.auth.service.AuthService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
